feat: enhance HymoFS configuration handling and logging
- Added logic to set HymoFS enabled state based on configuration during mount.
- Updated `set_enabled` method to use `int` for the enable parameter for consistency.
- Improved logging to provide clearer insights into HymoFS state changes.

+++ src/mount/hymofs.cpp
@@ -277,7 +277,7 @@ bool HymoFS::set_stealth(bool enable) {
 }
 
 bool HymoFS::set_enabled(bool enable) {
-    uint64_t val = enable ? 1 : 0;
+    int val = enable ? 1 : 0;
     LOG_INFO("HymoFS: Setting enabled=" + std::string(enable ? "true" : "false"));
     bool ret = hymo_execute_cmd(HYMO_IOC_SET_ENABLED, &val) == 0;
     if (!ret) {
